Switch to copying using target mechanism
In PK11_ExportEncryptedPrivKeyInfo, when trying to move the
generated/derived key to a different token, use the future mechanism
we'll be using this key under, not its generated mechanism. This
ensures, in scenarios when the new token doesn't do the original
mechanism but _does_ do the destination mech, the key is moved
correctly.

This behavior is observed with certian nCipher (now Entrust) HSMs (on
firmware version 12.60) in strict FIPS mode.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>